scripts/requirements.txt
openai-whisper
yt-dlp
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu121
